Renamed index and deleted unused files
